feat: Add comments count to Post component
Update the Post component to display the number of comments for each post. This is achieved by passing the 'comments' prop to the Post component and rendering the count in the link text. This change enhances the user experience by providing visibility into the number of comments on a post.