feat: Add queryClient.invalidateQueries(["posts"]) after successful post creation
This commit adds the queryClient.invalidateQueries(["posts"]) function call after a post is successfully created in the CreatePost component. This ensures that the "posts" query is invalidated and refetched, updating the list of posts displayed in the application. This change improves the data consistency and ensures that the newly created post is immediately visible to the user.